client: rename zoom plugin to camera plugin
